lession/sixth/requirements.txt

fastapi[all]
sqlalchemy
alembic
asyncpg
fastapi-users[sqlalchemy]
python-dotenv
fastapi-cache2[redis]
email-validator
celery
flower
